Functional & UI Improvements

+++ ajax.php
@@ -30,16 +30,20 @@ require_once('../../config.php');
  *
  * All permission-checks are done in the constructor of mod_confman_item
  */
+
 require_once($CFG->dirroot . '/mod/confman/lib.php');
-$PAGE->set_context(context_system::instance());
+
 
 $action = required_param("act", PARAM_ALPHANUMEXT);
 $itemid = required_param("id", PARAM_INT);
 $token = optional_param("token", "", PARAM_ALPHANUMEXT);
 
 $item = new mod_confman_item($itemid, $token);
+
+$PAGE->set_context(context_course::instance($item->event->course));
+
 // Now that we have created our item we check if we are allowed to access.
-if (!$item->can_edit && !$item->can_view) {
+if ($item->id == 0 || (!$item->can_edit && !$item->can_view)) {
     $OUTPUT->header();
     echo "<p>Permission denied</p>";
     echo $OUTPUT->footer();
@@ -50,23 +54,47 @@ $result = array();
 
 switch($action){
     case "file_append":
-        $filename = required_param("filename", PARAM_FILE);
-        $filecontent = required_param("file", PARAM_RAW);
-        $result["url"] = "".$item->file_append($filename, $filecontent);
-        if ($result["url"] != "") {
-            $result["status"] = "ok";
-        } else {
-            $result["status"] = "error";
+        if ($item->can_edit) {
+            $filename = required_param("filename", PARAM_TEXT);
+            $filecontent = required_param("file", PARAM_RAW);
+            $result["url"] = "".$item->file_append($filename, $filecontent);
+            if ($result["url"] != "") {
+                $result["status"] = "ok";
+            } else {
+                $result["status"] = "error";
+            }
         }
     break;
     case "file_delete":
-        $filename = required_param("filename", PARAM_ALPHANUMEXT);
-        $chk = $item->file_delete($filename);
-        $result["delete_file"] = $filename;
-        if ($chk) {
+        if ($item->can_edit) {
+            $filename = required_param("filename", PARAM_TEXT);
+            $chk = $item->file_delete($filename);
+            $result["delete_file"] = $filename;
+            if ($chk) {
+                $result["status"] = "ok";
+            } else {
+                $result["status"] = "error";
+            }
+        }
+    break;
+    case "file_mail":
+        if ($item->can_edit) {
+            $type = required_param("type", PARAM_TEXT);
+            if (!in_array($type, array("file_append", "file_delete"))) {
+                $result["status"] = "error";
+            } else {
+                $result["status"] = "ok";
+                $item->mail($type, 'files');
+            }
+        }
+    break;
+    case "set_approved":
+        if ($item->can_manage) {
+            $setto = optional_param('setto', 0, PARAM_INT);
+            $item->data->approved = $setto;
+            $item->store($item->data, 0);
+            $result["setto"] = $setto;
             $result["status"] = "ok";
-        } else {
-            $result["status"] = "error";
         }
     break;
 }
